third_party/WebKit/LayoutTests/fast/dom/shadow/nodetype.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../resources/js-test.js"></script>
</head>
<body>
<progress id="progress" max="100"></progress>
<pre id="console">
This tests the shadow host's and shadow root's nodeType.
</pre>
<script>
var progress = document.getElementById('progress');
var shadow = internals.shadowRoot(progress);
var shadowChild = shadow.firstChild;
shouldBe('progress.nodeType', 'Node.ELEMENT_NODE');
shouldBe('shadow.nodeType', 'Node.DOCUMENT_FRAGMENT_NODE');
shouldBeNonNull('shadowChild');
shouldBeTrue('shadowChild.nodeType != Node.DOCUMENT_FRAGMENT_NODE');
</script>
</body>
</html>
